fs: pstore: Fix a warning
Fixes: warning: format '%lx' expects argument of type 'long unsigned int', but argument 2 has type 'phys_addr_t' [-Wformat=]

+++ fs/pstore/ram.c
@@ -820,7 +820,7 @@ static int __init ramoops_memreserve(char *p)
 	ramoops_data.dump_oops = 1;
 
 	pr_info("msm_reserve_ramoops_memory addr=%lx,size=%lx\n",
-			ramoops_data.mem_address, ramoops_data.mem_size);
+			(unsigned long)ramoops_data.mem_address, ramoops_data.mem_size);
 	pr_info("msm_reserve_ramoops_memory record_size=%lx,ftrace_size=%lx\n",
 			ramoops_data.record_size, ramoops_data.ftrace_size);
